Instead of @auth, use @noauth (less error-prone); cleanup tests

+++ backend/tests/integration.py
@@ -31,27 +31,43 @@ def _delete_all_from_db():
         c.execute(sql)
     migrate_if_needed()
 
-
 def setup_module():
-    _delete_all_from_db()
-
+    pass
 
 def teardown_module():
     _delete_all_from_db()
 
-
 @pytest.fixture
 def app_client():
+    _delete_all_from_db()
     app.testing = True
     return app.test_client()
 
-def test_values_put_get(app_client):
+@pytest.fixture
+def first_admin_exists(app_client):
+    data = { 'name': 'First User - Admin', 'username': TEST_USERNAME, 'password': TEST_PASSWORD, 'email': 'test@example.com' }
+    r = app_client.post('/api/admin/first', data=json.dumps(data), content_type='application/json')
+    assert r.status_code == 201
+    admin_id = json.loads(r.data.decode('utf-8'))['id']
+    assert int(admin_id) == 1
+    return True
+
+@pytest.fixture
+def authorization_header(app_client, first_admin_exists):
+    data = { 'username': TEST_USERNAME, 'password': TEST_PASSWORD }
+    r = app_client.post('/api/auth/login', data=json.dumps(data), content_type='application/json')
+    assert r.status_code == 200
+    auth_header = dict(r.headers).get('X-JWT-Token', None)
+    assert auth_header[:9] == 'Bearer 1:'
+    return auth_header
+
+def test_values_put_get_simple(app_client, authorization_header):
     """
         Put a value, get a value.
     """
     data = [{'p': 'qqqq.wwww', 't': 1234567890.123456, 'v': 111.22}]
     app_client.put('/api/values/', data=json.dumps(data), content_type='application/json')
-    r = app_client.get('/api/values/?p=qqqq.wwww&t0=1234567890&t1=1234567891&a=no')
+    r = app_client.get('/api/values/?p=qqqq.wwww&t0=1234567890&t1=1234567891&a=no', headers={'Authorization': authorization_header})
     expected = {
         'paths': {
             'qqqq.wwww': {
@@ -62,17 +78,18 @@ def test_values_put_get(app_client):
             }
         }
     }
+    assert r.status_code == 200
     actual = json.loads(r.data.decode('utf-8'))
     assert expected == actual
 
-def test_values_put_get_noaggrparam_redirect(app_client):
+def test_values_put_get_noaggrparam_redirect(app_client, authorization_header):
     """
         Try to get values without aggr param, get redirected.
     """
     t_from = 1330000000
     t_to = 1330000000 + 100*15*60 + 1
     url = '/api/values/?p=aaaa.bbbb&t0={}&t1={}&max=10'.format(t_from, t_to)
-    r = app_client.get(url)
+    r = app_client.get(url, headers={'Authorization': authorization_header})
 
     assert r.status_code == 301
 
@@ -82,7 +99,7 @@ def test_values_put_get_noaggrparam_redirect(app_client):
             redirect_location = value
     assert redirect_location[-len(url)-5:] == url + '&a=no'
 
-def test_values_put_get_sort_limit(app_client):
+def test_values_put_get_sort_limit(app_client, authorization_header):
     """
         Try to get values without aggr param, get redirected.
     """
@@ -96,7 +113,7 @@ def test_values_put_get_sort_limit(app_client):
     app_client.put('/api/values/', data=json.dumps(data), content_type='application/json')
 
     url = '/api/values/?p={}&a=no&sort=desc&limit=2'.format(TEST_PATH)
-    r = app_client.get(url)
+    r = app_client.get(url, headers={'Authorization': authorization_header})
 
     expected = {
         'paths': {
@@ -115,7 +132,7 @@ def test_values_put_get_sort_limit(app_client):
     #pprint(actual)
     assert expected == actual
 
-def test_values_put_few_get_aggr(app_client):
+def test_values_put_few_get_aggr(app_client, authorization_header):
     """
         Put a few values, get aggregated value.
     """
@@ -131,7 +148,7 @@ def test_values_put_few_get_aggr(app_client):
     t_from = 1330002000  # aggr level 0 - every 1 hour
     t_to = t_from + 1*3600
     url = '/api/values/?p={}&t0={}&t1={}&a=0'.format(TEST_PATH, t_from, t_to)
-    r = app_client.get(url)
+    r = app_client.get(url, headers={'Authorization': authorization_header})
 
     expected = {
         'paths': {
@@ -150,7 +167,7 @@ def test_values_put_few_get_aggr(app_client):
     assert expected == actual
 
 #@pytest.mark.skip(reason="removing aggregated data is not implemented, so this test fails when repeated! Otherwise it works with fresh DB.")
-def test_values_put_many_get_aggr(app_client):
+def test_values_put_many_get_aggr(app_client, authorization_header):
     """
         Put many values, get aggregated value. Delete path and values.
     """
@@ -169,7 +186,7 @@ def test_values_put_many_get_aggr(app_client):
         app_client.put('/api/values/', data=json.dumps(data), content_type='application/json')
         #print(t_from, t_to)
         url = '/api/values/?p={}&t0={}&t1={}&max=10&a=3'.format(TEST_PATH, t_from, t_to)
-        r = app_client.get(url)
+        r = app_client.get(url, headers={'Authorization': authorization_header})
 
         expected = {
             'paths': {
@@ -187,10 +204,17 @@ def test_values_put_many_get_aggr(app_client):
         #pprint(actual)
         assert expected == actual
     finally:
-        app_client.delete('/api/paths/?p={}'.format(TEST_PATH))
+        r = app_client.delete('/api/paths/?p={}'.format(TEST_PATH), headers={'Authorization': authorization_header})
         assert r.status_code == 200
 
-def test_dashboards_widgets_post_get(app_client):
+def test_paths_delete_need_auth(app_client, authorization_header):
+    TEST_PATH = 'test.values.put.many.get.aggr'
+    r = app_client.delete('/api/paths/?p={}'.format(TEST_PATH))
+    assert r.status_code == 401
+    r = app_client.delete('/api/paths/?p={}'.format(TEST_PATH), headers={'Authorization': authorization_header})
+    assert r.status_code == 404  # because path is not found, of course
+
+def test_dashboards_widgets_post_get(app_client, authorization_header):
     """
         Create a dashboard, get a dashboard, create a widget, get a widget. Delete the dashboard, get 404.
     """
@@ -198,10 +222,10 @@ def test_dashboards_widgets_post_get(app_client):
     try:
         WIDGET = 'chart1'
         data = {'name': DASHBOARD + ' name', 'slug': DASHBOARD}
-        r = app_client.post('/api/dashboards/', data=json.dumps(data), content_type='application/json')
+        r = app_client.post('/api/dashboards/', data=json.dumps(data), content_type='application/json', headers={'Authorization': authorization_header})
         assert r.status_code == 201
 
-        r = app_client.get('/api/dashboards/{}'.format(DASHBOARD))
+        r = app_client.get('/api/dashboards/{}'.format(DASHBOARD), headers={'Authorization': authorization_header})
         expected = {
             'name': DASHBOARD + ' name',
             'slug': DASHBOARD,
@@ -223,11 +247,11 @@ def test_dashboards_widgets_post_get(app_client):
                 }
             ])
         }
-        r = app_client.post('/api/dashboards/{}/widgets/'.format(DASHBOARD), data=json.dumps(widget_post_data), content_type='application/json')
+        r = app_client.post('/api/dashboards/{}/widgets/'.format(DASHBOARD), data=json.dumps(widget_post_data), content_type='application/json', headers={'Authorization': authorization_header})
         assert r.status_code == 201
         widget_id = json.loads(r.data.decode('utf-8'))['id']
 
-        r = app_client.get('/api/dashboards/{}/widgets/'.format(DASHBOARD))
+        r = app_client.get('/api/dashboards/{}/widgets/'.format(DASHBOARD), headers={'Authorization': authorization_header})
         actual = json.loads(r.data.decode('utf-8'))
         widget_post_data['id'] = widget_id
         expected = {
@@ -251,11 +275,11 @@ def test_dashboards_widgets_post_get(app_client):
                 }
             ])
         }
-        r = app_client.put('/api/dashboards/{}/widgets/{}'.format(DASHBOARD, widget_id), data=json.dumps(widget_post_data), content_type='application/json')
+        r = app_client.put('/api/dashboards/{}/widgets/{}'.format(DASHBOARD, widget_id), data=json.dumps(widget_post_data), content_type='application/json', headers={'Authorization': authorization_header})
         assert r.status_code == 204
 
         # make sure it was updated:
-        r = app_client.get('/api/dashboards/{}/widgets/'.format(DASHBOARD))
+        r = app_client.get('/api/dashboards/{}/widgets/'.format(DASHBOARD), headers={'Authorization': authorization_header})
         actual = json.loads(r.data.decode('utf-8'))
         widget_post_data['id'] = widget_id
         expected = {
@@ -267,21 +291,21 @@ def test_dashboards_widgets_post_get(app_client):
         assert expected == actual
 
         # get a single widget:
-        r = app_client.get('/api/dashboards/{}/widgets/{}/'.format(DASHBOARD, widget_id))
+        r = app_client.get('/api/dashboards/{}/widgets/{}/'.format(DASHBOARD, widget_id), headers={'Authorization': authorization_header})
         actual = json.loads(r.data.decode('utf-8'))
         expected = widget_post_data
         assert r.status_code == 200
         assert expected == actual
 
         # delete dashboard:
-        app_client.delete('/api/dashboards/{}'.format(DASHBOARD))
-        r = app_client.get('/api/dashboards/{}'.format(DASHBOARD))
+        r = app_client.delete('/api/dashboards/{}'.format(DASHBOARD), headers={'Authorization': authorization_header})
+        assert r.status_code == 200
+        r = app_client.get('/api/dashboards/{}'.format(DASHBOARD), headers={'Authorization': authorization_header})
         assert r.status_code == 404
 
     except:
         # if something went wrong, delete dashboard so the next run can succeed:
         app_client.delete('/api/dashboards/{}'.format(DASHBOARD))
-        r = app_client.get('/api/dashboards/{}'.format(DASHBOARD))
         raise
 
 @pytest.mark.skip(reason="no idea.")
@@ -362,7 +386,7 @@ def test_accounts(app_client):
     assert r.status_code == 200
 
 
-def test_jwt_expiry_refresh(app_client):
+def test_jwt_expiry_refresh(app_client, first_admin_exists):
     """
         Login, get X-JWT-Token which expired before 1s, make sure you get 401 on resource request,
         hit /auth/refrest, make sure the new token works
@@ -397,7 +421,7 @@ def test_jwt_expiry_refresh(app_client):
     assert r.status_code == 200
 
 
-def test_jwt_total_expiry(app_client):
+def test_jwt_total_expiry(app_client, first_admin_exists):
     """
         Login, get X-JWT-Token which expired before 1s, read the new token from header, check it
         WARNING: depends on test_accounts() being run first
